# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Home Metrics API
  version: 1.0.0
  description: REST API served by hm-api-server for sensor, energy, and collector status data.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiToken: []
paths:
  /api/health:
    get:
      summary: Health check
      security: []
      responses:
        "200":
          description: Database is reachable
  /api/health/details:
    get:
      summary: Collector health summary
      responses:
        "200":
          description: Collector summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthDetails"
  /api/devices:
    get:
      summary: List sensor devices
      responses:
        "200":
          description: Sensor devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
  /api/devices/{mac}/latest:
    get:
      summary: Latest sensor values
      parameters:
        - $ref: "#/components/parameters/Mac"
      responses:
        "200":
          description: Latest values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestSensor"
  /api/devices/{mac}/series:
    get:
      summary: Sensor series
      parameters:
        - $ref: "#/components/parameters/Mac"
        - name: metric
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SensorMetric"
        - name: range
          in: query
          schema:
            $ref: "#/components/schemas/Range"
      responses:
        "200":
          description: Sensor series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorSeries"
  /api/energy/latest:
    get:
      summary: Latest energy values
      parameters:
        - name: source
          in: query
          schema:
            type: string
        - name: device_key
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Energy latest readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnergyLatest"
  /api/energy/series:
    get:
      summary: Energy series
      parameters:
        - name: source
          in: query
          required: true
          schema:
            type: string
        - name: device_key
          in: query
          required: true
          schema:
            type: string
        - name: metric
          in: query
          required: true
          schema:
            type: string
        - name: range
          in: query
          schema:
            $ref: "#/components/schemas/Range"
      responses:
        "200":
          description: Energy series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnergySeries"
  /api/admin/collector-status:
    get:
      summary: Collector status rows
      responses:
        "200":
          description: Collector status rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectorStatus"
  /api/admin/cisco-spaces-firehose:
    get:
      summary: Cisco Spaces firehose collector status
      responses:
        "200":
          description: Cisco Spaces firehose status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CiscoSpacesFirehoseStatus"
  /api/admin/schema:
    get:
      summary: DB migration status
      responses:
        "200":
          description: Applied migrations
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiToken:
      type: apiKey
      in: header
      name: X-API-Token
  parameters:
    Mac:
      name: mac
      in: path
      required: true
      schema:
        type: string
  schemas:
    Range:
      type: string
      enum: [1d, 1w, 1m, 3m, 1y]
      default: 1d
    SensorMetric:
      type: string
      enum:
        - temperature_c
        - humidity_percent
        - battery_percent
        - rssi_dbm
        - pressure_hpa
        - co2_ppm
        - lux
        - etvoc
    Device:
      type: object
      required: [mac, label, enabled]
      properties:
        mac:
          type: string
        label:
          type: string
        sensor_category:
          type: string
        location:
          type: string
        enabled:
          type: boolean
    LatestSensor:
      type: object
      required: [device, ts, values]
      properties:
        device:
          $ref: "#/components/schemas/Device"
        ts:
          type: string
          format: date-time
        values:
          type: object
          additionalProperties:
            type:
              - number
              - "null"
    SeriesPoint:
      type: object
      required: [ts, value]
      properties:
        ts:
          type: string
          format: date-time
        value:
          type: number
    SensorSeries:
      type: object
      required: [mac, metric, range, points]
      properties:
        mac:
          type: string
        metric:
          $ref: "#/components/schemas/SensorMetric"
        range:
          $ref: "#/components/schemas/Range"
        points:
          type: array
          items:
            $ref: "#/components/schemas/SeriesPoint"
    EnergyLatest:
      type: object
      required: [ts, source, device_key, metric, value]
      properties:
        ts:
          type: string
          format: date-time
        source:
          type: string
        device_key:
          type: string
        label:
          type: string
        location:
          type: string
        metric:
          type: string
        value:
          type: number
        unit:
          type: string
    EnergySeries:
      type: object
      required: [source, device_key, metric, range, unit, points]
      properties:
        source:
          type: string
        device_key:
          type: string
        metric:
          type: string
        range:
          $ref: "#/components/schemas/Range"
        unit:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/SeriesPoint"
    HealthDetails:
      type: object
      required: [status, database, collector_targets, stale_collectors]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        database:
          type: string
        collector_targets:
          type: integer
        stale_collectors:
          type: integer
    CollectorStatus:
      type: object
      required: [collector_name, target_type, target_key, consecutive_failures, updated_at]
      properties:
        collector_name:
          type: string
        target_type:
          type: string
        target_key:
          type: string
        last_attempt_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        last_data_at:
          type: string
          format: date-time
        first_failure_at:
          type: string
          format: date-time
        last_failure_at:
          type: string
          format: date-time
        last_error:
          type: string
        consecutive_failures:
          type: integer
        updated_at:
          type: string
          format: date-time
    CiscoSpacesFirehoseStatus:
      type: object
      required:
        - lock_key
        - lock_held
        - mode
        - configured_secondary_allowed
        - collector_status_present
        - collector_reporting
        - consecutive_failures
      properties:
        lock_key:
          type: integer
        lock_held:
          type: boolean
        lock_holder_pid:
          type: integer
        mode:
          type: string
        configured_secondary_allowed:
          type: boolean
        collector_status_present:
          type: boolean
        collector_reporting:
          type: boolean
        collector_updated_at:
          type: string
          format: date-time
        last_attempt_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        last_data_at:
          type: string
          format: date-time
        first_failure_at:
          type: string
          format: date-time
        last_failure_at:
          type: string
          format: date-time
        consecutive_failures:
          type: integer
        last_error:
          type: string